requirements.txt
altgraph==0.16.1
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
codecov==2.0.15
coverage==4.5.3
coveralls==1.7.0
docopt==0.6.2
future==0.17.1
idna==2.8
lectuure-downloader==0.1
lxml==4.3.3
macholib==1.11
more-itertools==7.0.0
pathlib2==2.3.3
pefile==2019.4.18
pluggy==0.9.0
py==1.8.0
pycparser==2.19
PyInstaller==3.4
pytest==4.4.1
pytest-cov==2.6.1
pytest-ordering==0.6
PyYAML==5.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.24.2
wget==3.2
